pyproject.toml
[project]
name = "nvidia-nim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-nvidia-ai-endpoints>=0.3.7",
    "python-dotenv>=1.0.1",
]
